remove buffer switching binds
The buffer switching binds are so quick and easy that I end up using
them for all my buffer switching, but when there are too many buffers
I lose orientation. So I am dropping them to force myself to reach for
qbuf instead, so I will get into the habit of using it.